refactor: Summary 파이프라인 결과를 summary 패키지 아래로 변경
